+++ methods/src/test/kotlin/methods/equations/DecimalLinearEquationsTest.kt
@@ -53,54 +53,6 @@ class DecimalLinearEquationsTest {
     }
 
     @Test
-    fun `test decimal linear equation with solution not expressible as terminating decimal`() = testMethodInX {
-        method = EquationsPlans.SolveDecimalLinearEquation
-        inputExpr = "3.1 x + 2.2 = 1.21"
-
-        check {
-            fromExpr = "3.1 x + 2.2 = 1.21"
-            toExpr = "SetSolution[x: {-[99 / 310]}]"
-            explanation {
-                key = EquationsExplanation.SolveDecimalLinearEquation
-            }
-
-            step {
-                fromExpr = "3.1 x + 2.2 = 1.21"
-                toExpr = "3.1 x = -0.99"
-                explanation {
-                    key = methods.solvable.EquationsExplanation.MoveConstantsToTheRightAndSimplify
-                }
-            }
-
-            step {
-                fromExpr = "3.1 x = -0.99"
-                toExpr = "x = -[99 / 310]"
-                explanation {
-                    key = methods.solvable.EquationsExplanation.DivideByCoefficientOfVariableAndSimplify
-                }
-            }
-
-            step {
-                fromExpr = "x = -[99 / 310]"
-                toExpr = "SetSolution[x: {-[99 / 310]}]"
-                explanation {
-                    key = EquationsExplanation.ExtractSolutionFromEquationInSolvedForm
-                }
-            }
-        }
-    }
-
-    @Test
-    fun `test decimal linear equation fails for recurring decimal in solution`() = testMethodInX {
-        method = EquationsPlans.SolveDecimalLinearEquation
-        inputExpr = "3.1x + 2.2[3] = 1.21"
-
-        check {
-            noTransformation()
-        }
-    }
-
-    @Test
     fun `test decimal linear equation with fractions in the initial expression`() = testMethodInX {
         method = EquationsPlans.SolveDecimalLinearEquation
         inputExpr = "3.6 x + 2.2 = [2 / 5] x + 1.2"